grrr

+++ src/optimizers/adamw.py
@@ -112,12 +112,12 @@ class AdamW(Optimizer):
                     bias_correction2 = 1.0 - beta2 ** state["step"]
                     step_size = step_size * math.sqrt(bias_correction2) / bias_correction1
 
-                update = exp_avg / denom
+                update = (exp_avg / denom).to(p.dtype)
                 update = torch.nan_to_num(update, nan=0.0, posinf=0.0, neginf=0.0)
                 if group["update_clip"] is not None:
                     update = torch.clamp(update, -group["update_clip"], group["update_clip"])
 
-                p.add_(update.to(p.dtype), alpha=-step_size)
+                p.add_(update, alpha=-step_size)
 
                 # Just adding the square of the weights to the loss function is *not*
                 # the correct way of using L2 regularization/weight decay with Adam,
